Steemit A11y - outline: none on all links (#278)
Removing outlines on links hampers the accessibility of the site for people with impaired vision etc. Thus it is important not to disable outlines as they are there for a reason.
I'm creating this pull-request at the same time as I'm publishing a post on steemit with a list of changes that should be implemented to try and make steemit as accessible as possible, so sadly it isn't of the highest quality. I'm making this mostly as a suggestion and to raise awareness.