remove ReplyEditor__title padding
this was brought in by 647bcfe8 and breaks the title input
by hiding text (in firefox 49.0.2).

surprisingly does not break things in chrome/opera, making
this a browser quirk. also, padding on input elements should
not really be necessary, i guess. browsers will take care of
things on its own.

Signed-off-by: thewisenerd <thewisenerd@protonmail.com>