de-frag translations
closes #2349

fixes some of #2351 but that will need more work

still need to figure out a way to interpolate react dom elements into a
translation.

also, this commit adds a storybook to cover all cases for
TransferHistoryRow.

this story provides knobs for any incoming prop that affects branching
and can be used as the basis for a real test eventually